requirements.txt
fastapi
uvicorn
python-multipart
requests
chromadb
sentence-transformers
openai-whisper
ollama
jinja2
fpdf
pypdf
numpy